[OC-4] Modify code based on PR comments
